10_여명구/03_OutputParser/requirements.txt
python-dotenv==1.0.1
langchain==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
pydantic==2.10.5